CPU, run command, PCB display, CPU display
Created a CPU and PCB display to display the state of the PCB and the
CPU as a process is running and terminated. Added these functions to
the virtualServices. Created the CPU: updated the cycle function so it
executes the current PCB and updates the CPU display. created execture
function to execute the next relocation value OPCODE. Created a suite
of functions to manage the assembly 6502 opcodes in spec. Also
redesigned the page a little in general just to make everything more
viewable. This commit is long because I worked on it in donnelly on my
laptop and did not have access to the internet on my laptop to commit
more frequently. I also fixed my bsod command from project 1.